prod-requirements.txt
asgiref==3.2.10
Django==3.1.14
django-allauth==0.43.0
djangorestframework==3.12.1
django_rest_swagger==2.2.0
django-filter==2.4.0
django-typomatic==1.5.0
pdf2image==1.14.0
Pillow==9.0.1
python-magic==0.4.18
pytz==2020.1
sqlparse==0.3.1
drf-writable-nested==0.6.3
